physical-electron-density-vertical-profile
In this Python code, we impose the physical electron density profile as a function of time and location.
We inherited the methodology presented in NequickG package which is supplied by the European Union Agency for the Space Program.
Source: https://www.gsc-europa.eu/support-to-developers/ionospheric-correction-algorithms/nequick-g-source-code

Some reviews were covered to reach a more understanding of coding parameters, as published in https://hamsci.org/publications/hamsci-distributed-array-small-instruments-personal-space-weather-station-dasi-psws.

Our code hinges on the knowing of three main parameters, latitude, longitude and time.
NequickG is used to set the used parameters as follows:
NEQTime: The time when the ionosphere electron density is modeled in Year, Month and Hour.
Position: Set by the known latitude and longitude.
resolution_lat: The latitude resolution, initially set to be 2.5 degrees.
resolution_lon: The longitude resolution, initially set to be 5 degrees.
required_height_of_study: The required height of study interest, initially set to be  400 Kms

GalileoBroadcast: This is the coefficients reference from which the physical model is fed with. For instance, the coefficients are as follows:
## Gallileo Coefficients backing to January 1st, 2017
a0 = 4.4000e+01
a1 = 3.8281e-01
a2 = -1.8616e-03
Source: National Aeronautics and Space Administration: https://cddis.nasa.gov/archive/gps/data/daily/2019/brdc/
  
The output is shown in a .csv file that comprises: Latitudes; Longitudes and Electron Density.

Usage:
Run the main Python script "physical-electron-density-vertical-profile" which calls the required classes.