MATLAB Program to get the Electron density:
CDF format files are in a NASA SPDF format. So, we need to download a MATLAB patch from NASA SPDF website to read the ISS data files. 
The MATLAB patch has to be downloaded from here:
https://cdf.gsfc.nasa.gov/html/matlab_cdf_patch.html

The link to download the ISS electron density Data:
https://spdf.gsfc.nasa.gov/pub/data/international_space_station_iss/sp_fpmu/

Yes, now you are ready to run the program.